% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{app_server}
\usage{
app_server(input, output)
}
\arguments{
\item{input}{choice of the shinyapp user}

\item{output}{result of these choices}
}
\value{
functions for the shinyapp
}
\description{
app_server
}
\examples{
# appserver(input,output)
}
